Add images to a estate Get type for a estate

+++ app/src/main/java/com/openclassrooms/realestatemanager/data/dao/EstateImageDao.kt
@@ -7,7 +7,7 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface EstateImageDao {
     @Query("SELECT * FROM estate_image WHERE estate_id = :estateId ORDER BY id")
-    fun getImagesForAEstate(estateId: Int?): Flow<List<EstateImage>>
+    fun getImagesForAEstate(estateId: Long?): Flow<List<EstateImage>>
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
     suspend fun insert(estateImage: EstateImage)
